chore: refactor the codebase
