+++ server/controllers/auth.js
@@ -12,44 +12,12 @@ const resetUrl = process.env.RESET_URL;
  */
 export default class AuthController {
   /**
-   * @method registerUser
-   * @description registers a user with their email
-   * @param {*} req
-   * @param {*} res
-   * @returns {object} registered user
-   */
-  static async signUpController(req, res) {
-    const { body } = req;
-
-    const userEmail = body.email;
-    body.email = userEmail.toLowerCase();
-
-    try {
-      UserModel.create(body, (err, data) => {
-        const { _id, email, isAdmin } = data;
-        const token = authHelper.encode({ _id, email, isAdmin });
-
-        res.status(200).json({
-          success: true,
-          token,
-          data,
-        });
-      });
-    } catch (error) {
-      return res.status(500).send({
-        success: false,
-        error,
-      });
-    }
-  }
-  /**
    * @method loginUser
    * @description logs in a user with their email and password
    * @param {*} req
    * @param {*} res
    * @returns {object} logged in user without parcel property
    */
-
   static async loginController(req, res) {
     const { email, password } = req.body;
     const convertedEmail = email.toLowerCase();
@@ -81,6 +49,7 @@ export default class AuthController {
       user,
     });
   }
+
   /**
    * @method passwordResetRequest
    * @description creates unique password reset link using user id and email to create token
@@ -88,7 +57,6 @@ export default class AuthController {
    * @param {*} res
    * @returns {object} token created
    */
-
   static async passwordResetRequest(req, res) {
     try {
       const { email } = req.body;
@@ -128,15 +96,15 @@ export default class AuthController {
       });
     }
   }
+
   /**
-   * @method password reset page request
+   * @method passwordResetPageRequest
    * @description validate password reset token  and sends a result on success or not
    * @param {*} req
    * @param {*} res
    * @route GET api/v1/auth/reset/:token
-   * @returns success message
+   * @returns {object} success message
    */
-
   static async passwordResetPageRequest(req, res) {
     try {
       const user = await UserModel.findOne({
@@ -160,15 +128,15 @@ export default class AuthController {
       });
     }
   }
+
   /**
-   * @method reset password
+   * @method updatePassword
    * @description validate password reset token and updates user password with hashed new password. then sets the token property to undefined
    * @param {*} req
    * @param {*} res
    * @route POST api/v1/auth/reset/:token
    * @returns {object} user with hashed updated password
    */
-
   static async updatePassword(req, res) {
     const user = await UserModel.findOne({
       resetPasswordToken: req.params.token,
@@ -199,14 +167,14 @@ export default class AuthController {
       });
     }
   }
+
   /**
-   * @method social redirect
+   * @method socialRedirect
    * @description resolves the social login route.
    * @param {*} req
    * @param {*} res
    * @returns {object} success message with social user firstname and lastname
    */
-
   static async socialRedirect(req, res) {
     const { firstName, lastName } = req.user;
     return res.status(200).json({
@@ -214,13 +182,16 @@ export default class AuthController {
       message: `Social user ${firstName} ${lastName} is logged in`,
     });
   }
-  /**
-   * @method google social login
-   * @description checks if google user exists in app or not. create new user if not
-   * @param {*} req
-   * @param {*} res
-   */
 
+  /**
+   * @method googlePassportCallback
+   * @description checks if google user exists in app or not. create new user if not
+   * @param {*} accessToken
+   * @param {*} refreshToken
+   * @param {*} profile
+   * @param {*} done
+   * @returns {Promise} a done function that accepts null and user object
+   */
   static async googlePassportCallback(
     accessToken,
     refreshToken,
@@ -252,13 +223,16 @@ export default class AuthController {
       }
     );
   }
-  /**
-   * @method github social login
-   * @description checks if githubuser exists in app or not. create new user if not
-   * @param {*} req
-   * @param {*} res
-   */
 
+  /**
+   * @method githubPassportCallback
+   * @description checks if githubuser exists in app or not. create new user if not
+   * @param {*} accessToken
+   * @param {*} refreshToken
+   * @param {*} profile
+   * @param {*} done
+   * @returns {Promise} a done function that accepts null and user object
+   */
   static async githubPassportCallback(
     accessToken,
     refreshToken,
